doctor,book service,book plan Trash work done

+++ resources/views/admin/manageDoctorTrash.blade.php
@@ -1,31 +1,30 @@
 @extends('admin.adminBase')
 
 @section('content')
+    <div class="flex-1 flex mt-12 items-center justify-between ">
+        <h1 class="text-lg font-semibold  py-2">Manage Doctors Trash (<span id="counting">0</span>)</h1>
+        <a href="/admin/manage-doctor/insert" class="bg-indigo-500 hover:bg-indigo-800 text-white px-3 py-2 rounded">
+            <i class="fas fa-plus"></i> Add New Doctor</a>
 
-<div class="flex-1 flex mt-12 items-center justify-between ">
-    <h1 class="text-lg font-semibold  py-2">Manage Doctors (<span id="counting">0</span>)</h1>
-    <a href="/admin/manage-doctor/insert" class="bg-indigo-500 hover:bg-indigo-800 text-white px-3 py-2 rounded">
-        <i class="fas fa-plus"></i> Add New Doctor</a>
-
-</div>
-<div class="overflow-x-auto">
-    <div class="relative overflow-x-auto shadow-md sm:rounded-lg">
-        <div class="pb-4">
-            <label for="table-search" class="sr-only">Search</label>
-            <div class="relative mt-1">
-                <div class="absolute inset-y-0 rtl:inset-r-0 start-0 flex items-center ps-3 pointer-events-none">
-                    <svg class="w-4 h-4 text-gray-500 dark:text-gray-400" aria-hidden="true"
-                        xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 20 20">
-                        <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
-                            d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z" />
-                    </svg>
+    </div>
+    <div class="overflow-x-auto">
+        <div class="relative overflow-x-auto shadow-md sm:rounded-lg">
+            <div class="pb-4">
+                <label for="table-search" class="sr-only">Search</label>
+                <div class="relative mt-1">
+                    <div class="absolute inset-y-0 rtl:inset-r-0 start-0 flex items-center ps-3 pointer-events-none">
+                        <svg class="w-4 h-4 text-gray-500 dark:text-gray-400" aria-hidden="true"
+                            xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 20 20">
+                            <path stroke="currentColor" stroke-linecap="round" stroke-linejoin="round" stroke-width="2"
+                                d="m19 19-4-4m0-7A7 7 0 1 1 1 8a7 7 0 0 1 14 0Z" />
+                        </svg>
+                    </div>
+                    <input type="text" name="query" id="searchInput"
+                        class="block pt-2 ps-10 text-sm text-gray-900 border border-gray-300 rounded-lg w-80 bg-gray-50 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
+                        placeholder="Search Our Doctor">
                 </div>
-                <input type="text" name="query" id="searchInput"
-                    class="block pt-2 ps-10 text-sm text-gray-900 border border-gray-300 rounded-lg w-80 bg-gray-50 focus:ring-blue-500 focus:border-blue-500 dark:bg-gray-700 dark:border-gray-600 dark:placeholder-gray-400 dark:text-white dark:focus:ring-blue-500 dark:focus:border-blue-500"
-                    placeholder="Search Our Doctor">
             </div>
-        </div>
-        <table class="min-w-full bg-white border border-gray-200">
+            <table class="min-w-full bg-white border border-gray-200">
                 <thead class="bg-gray-100">
                     <tr>
                         <th class="border-b border-gray-200 px-3 py-2 text-sm">Id</th>
@@ -38,31 +37,7 @@
                         <th class="border-b border-gray-200 px-3 py-2 text-sm">Actions</th>
                     </tr>
                 </thead>
-                <tbody id="callingDoctor" >
-                    <!-- <tr class="mt-5 capitalize">
-                        <td class="border-b border-gray-200 px-3 text-center py-2 text-sm">3</td> 
-                        <td class="border-b border-gray-200 px-3 text-center py-2 text-sm">3</td> 
-                        <td class="border-b border-gray-200 px-3 text-center py-2 text-sm">Dr. Rajiv Kumar</td> 
-                        <td class="border-b border-gray-200 px-3 text-center py-2 text-sm">Dentist</td> 
-                        <td class="border-b border-gray-200 px-3 text-center py-2 text-sm">Line Bazar, Purnea</td> 
-                        <td class="border-b border-gray-200 px-3 text-center py-2 text-sm">894561236</td> 
-
-                        <td class="border-b border-gray-200 px-3 text-center py-2 text-sm flex"> 
-                            {{-- <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512"class="h-3 mt-1 mr-2 "><path d="M0 64C0 46.3 14.3 32 32 32H96h16H288c17.7 0 32 14.3 32 32s-14.3 32-32 32H231.8c9.6 14.4 16.7 30.6 20.7 48H288c17.7 0 32 14.3 32 32s-14.3 32-32 32H252.4c-13.2 58.3-61.9 103.2-122.2 110.9L274.6 422c14.4 10.3 17.7 30.3 7.4 44.6s-30.3 17.7-44.6 7.4L13.4 314C2.1 306-2.7 291.5 1.5 278.2S18.1 256 32 256h80c32.8 0 61-19.7 73.3-48H32c-17.7 0-32-14.3-32-32s14.3-32 32-32H185.3C173 115.7 144.8 96 112 96H96 32C14.3 96 0 81.7 0 64z"/></svg>500 <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512" class="h-5 ml-2"><path d="M160 48a48 48 0 1 1 96 0 48 48 0 1 1 -96 0zM126.5 199.3c-1 .4-1.9 .8-2.9 1.2l-8 3.5c-16.4 7.3-29 21.2-34.7 38.2l-2.6 7.8c-5.6 16.8-23.7 25.8-40.5 20.2s-25.8-23.7-20.2-40.5l2.6-7.8c11.4-34.1 36.6-61.9 69.4-76.5l8-3.5c20.8-9.2 43.3-14 66.1-14c44.6 0 84.8 26.8 101.9 67.9L281 232.7l21.4 10.7c15.8 7.9 22.2 27.1 14.3 42.9s-27.1 22.2-42.9 14.3L247 287.3c-10.3-5.2-18.4-13.8-22.8-24.5l-9.6-23-19.3 65.5 49.5 54c5.4 5.9 9.2 13 11.2 20.8l23 92.1c4.3 17.1-6.1 34.5-23.3 38.8s-34.5-6.1-38.8-23.3l-22-88.1-70.7-77.1c-14.8-16.1-20.3-38.6-14.7-59.7l16.9-63.5zM68.7 398l25-62.4c2.1 3 4.5 5.8 7 8.6l40.7 44.4-14.5 36.2c-2.4 6-6 11.5-10.6 16.1L54.6 502.6c-12.5 12.5-32.8 12.5-45.3 0s-12.5-32.8 0-45.3L68.7 398z"/></svg>--}}
-
-                            <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512"class="h-3 mt-1 mr-2 "><path d="M0 64C0 46.3 14.3 32 32 32H96h16H288c17.7 0 32 14.3 32 32s-14.3 32-32 32H231.8c9.6 14.4 16.7 30.6 20.7 48H288c17.7 0 32 14.3 32 32s-14.3 32-32 32H252.4c-13.2 58.3-61.9 103.2-122.2 110.9L274.6 422c14.4 10.3 17.7 30.3 7.4 44.6s-30.3 17.7-44.6 7.4L13.4 314C2.1 306-2.7 291.5 1.5 278.2S18.1 256 32 256h80c32.8 0 61-19.7 73.3-48H32c-17.7 0-32-14.3-32-32s14.3-32 32-32H185.3C173 115.7 144.8 96 112 96H96 32C14.3 96 0 81.7 0 64z"/></svg>300<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 576 512" class="h-4 ml-2"><path d="M0 128C0 92.7 28.7 64 64 64H320c35.3 0 64 28.7 64 64V384c0 35.3-28.7 64-64 64H64c-35.3 0-64-28.7-64-64V128zM559.1 99.8c10.4 5.6 16.9 16.4 16.9 28.2V384c0 11.8-6.5 22.6-16.9 28.2s-23 5-32.9-1.6l-96-64L416 337.1V320 192 174.9l14.2-9.5 96-64c9.8-6.5 22.4-7.2 32.9-1.6z"/></svg> 
-                        </td>          
-
-
-                        <td class="border-b border-gray-200 px-3 text-center py-2 text-sm">
-                            {{--<button class="mx-8 text-sm"><svg xmlns="http://www.w3.org/2000/svg"class="h-5" viewBox="0 0 448 512"><path d="M0 96C0 78.3 14.3 64 32 64H416c17.7 0 32 14.3 32 32s-14.3 32-32 32H32C14.3 128 0 113.7 0 96zM0 256c0-17.7 14.3-32 32-32H416c17.7 0 32 14.3 32 32s-14.3 32-32 32H32c-17.7 0-32-14.3-32-32zM448 416c0 17.7-14.3 32-32 32H32c-17.7 0-32-14.3-32-32s14.3-32 32-32H416c17.7 0 32 14.3 32 32z"/></svg></button>--}}
-
-                            <button class=" py-1 px-2  "><svg class="w-5 h-5 text-gray-500 transition duration-75 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white" xmlns="http://www.w3.org/2000/svg" fill="currentColor" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 576 512"><path d="M288 32c-80.8 0-145.5 36.8-192.6 80.6C48.6 156 17.3 208 2.5 243.7c-3.3 7.9-3.3 16.7 0 24.6C17.3 304 48.6 356 95.4 399.4C142.5 443.2 207.2 480 288 480s145.5-36.8 192.6-80.6c46.8-43.5 78.1-95.4 93-131.1c3.3-7.9 3.3-16.7 0-24.6c-14.9-35.7-46.2-87.7-93-131.1C433.5 68.8 368.8 32 288 32zM144 256a144 144 0 1 1 288 0 144 144 0 1 1 -288 0zm144-64c0 35.3-28.7 64-64 64c-7.1 0-13.9-1.2-20.3-3.3c-5.5-1.8-11.9 1.6-11.7 7.4c.3 6.9 1.3 13.8 3.2 20.7c13.7 51.2 66.4 81.6 117.6 67.9s81.6-66.4 67.9-117.6c-11.1-41.5-47.8-69.4-88.6-71.1c-5.8-.2-9.2 6.1-7.4 11.7c2.1 6.4 3.3 13.2 3.3 20.3z"/></svg></button>
-                            <button class=" py-1 px-2  "><svg class="w-5 h-5 text-gray-500 transition duration-75 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 512 512"><path d="M471.6 21.7c-21.9-21.9-57.3-21.9-79.2 0L362.3 51.7l97.9 97.9 30.1-30.1c21.9-21.9 21.9-57.3 0-79.2L471.6 21.7zm-299.2 220c-6.1 6.1-10.8 13.6-13.5 21.9l-29.6 88.8c-2.9 8.6-.6 18.1 5.8 24.6s15.9 8.7 24.6 5.8l88.8-29.6c8.2-2.7 15.7-7.4 21.9-13.5L437.7 172.3 339.7 74.3 172.4 241.7zM96 64C43 64 0 107 0 160V416c0 53 43 96 96 96H352c53 0 96-43 96-96V320c0-17.7-14.3-32-32-32s-32 14.3-32 32v96c0 17.7-14.3 32-32 32H96c-17.7 0-32-14.3-32-32V160c0-17.7 14.3-32 32-32h96c17.7 0 32-14.3 32-32s-14.3-32-32-32H96z"/></svg></button>
-                            <button type='button' class='py-1 px-2  '><svg  class="w-5 h-5 text-gray-500 transition duration-75 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white" fill="currentColor"  xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path d="M135.2 17.7L128 32H32C14.3 32 0 46.3 0 64S14.3 96 32 96H416c17.7 0 32-14.3 32-32s-14.3-32-32-32H320l-7.2-14.3C307.4 6.8 296.3 0 284.2 0H163.8c-12.1 0-23.2 6.8-28.6 17.7zM416 128H32L53.2 467c1.6 25.3 22.6 45 47.9 45H346.9c25.3 0 46.3-19.7 47.9-45L416 128z"/></svg></button>
-                        </td>
-                    </tr>
-                   -->
+                <tbody id="callingDoctor">
                 </tbody>
                 <tfoot>
                     <tr>
@@ -75,24 +50,24 @@
         </div>
     </div>
 
-<script>
-    $(document).ready(function() {
-        // Function to fetch and display doctor
-        let callingDoctor = () => {
-            $.ajax({
-                type: "GET",
-                url: "{{ route('doctor.index') }}",
-                success: function(response) {
-                    let table = $("#callingDoctor");
-                    table.empty();
-                    let data = response.data;
+    <script>
+        $(document).ready(function() {
+            // Function to fetch and display doctor
+            let callingDoctor = () => {
+                $.ajax({
+                    type: "GET",
+                    url: "{{ route('doctor.index.deleted') }}",
+                    success: function(response) {
+                        let table = $("#callingDoctor");
+                        table.empty();
+                        let data = response.data;
 
-                    // Update doctor count
-                    let len = data.length;
-                    $("#counting").html(len);
+                        // Update doctor count
+                        let len = data.length;
+                        $("#counting").html(len);
 
-                    data.forEach((data) => {
-                        table.append(`
+                        data.forEach((data) => {
+                            table.append(`
                             <tr class="hover:bg-gray-100 mt-5 px-3 py-3 capitalize">
                                 <td class="border-b border-gray-200 text-center text-sm">${data.id}</td> 
                                 <td class="border-b border-gray-200 text-center text-sm "><img src="/image/doctor/${data.image}" width="40px" height="20px" alt=""></td> 
@@ -106,49 +81,69 @@
                             <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 320 512"class="h-3 mt-1 mr-2 "><path d="M0 64C0 46.3 14.3 32 32 32H96h16H288c17.7 0 32 14.3 32 32s-14.3 32-32 32H231.8c9.6 14.4 16.7 30.6 20.7 48H288c17.7 0 32 14.3 32 32s-14.3 32-32 32H252.4c-13.2 58.3-61.9 103.2-122.2 110.9L274.6 422c14.4 10.3 17.7 30.3 7.4 44.6s-30.3 17.7-44.6 7.4L13.4 314C2.1 306-2.7 291.5 1.5 278.2S18.1 256 32 256h80c32.8 0 61-19.7 73.3-48H32c-17.7 0-32-14.3-32-32s14.3-32 32-32H185.3C173 115.7 144.8 96 112 96H96 32C14.3 96 0 81.7 0 64z"/></svg>${data.online_charge}<svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 576 512" class="h-4 ml-2"><path d="M0 128C0 92.7 28.7 64 64 64H320c35.3 0 64 28.7 64 64V384c0 35.3-28.7 64-64 64H64c-35.3 0-64-28.7-64-64V128zM559.1 99.8c10.4 5.6 16.9 16.4 16.9 28.2V384c0 11.8-6.5 22.6-16.9 28.2s-23 5-32.9-1.6l-96-64L416 337.1V320 192 174.9l14.2-9.5 96-64c9.8-6.5 22.4-7.2 32.9-1.6z"/></svg> 
                         </td>          
                         <td class="border-b border-gray-200 px-3 text-center py-2 text-sm">
-                            <a href='/admin/manage-doctor/view/${data.id}'>
-                            <button class=" py-1 px-2 "><svg class="w-5 h-5 text-gray-500 transition duration-75 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white" xmlns="http://www.w3.org/2000/svg" fill="currentColor" xmlns="http://www.w3.org/2000/svg" viewBox="0 0 576 512"><path d="M288 32c-80.8 0-145.5 36.8-192.6 80.6C48.6 156 17.3 208 2.5 243.7c-3.3 7.9-3.3 16.7 0 24.6C17.3 304 48.6 356 95.4 399.4C142.5 443.2 207.2 480 288 480s145.5-36.8 192.6-80.6c46.8-43.5 78.1-95.4 93-131.1c3.3-7.9 3.3-16.7 0-24.6c-14.9-35.7-46.2-87.7-93-131.1C433.5 68.8 368.8 32 288 32zM144 256a144 144 0 1 1 288 0 144 144 0 1 1 -288 0zm144-64c0 35.3-28.7 64-64 64c-7.1 0-13.9-1.2-20.3-3.3c-5.5-1.8-11.9 1.6-11.7 7.4c.3 6.9 1.3 13.8 3.2 20.7c13.7 51.2 66.4 81.6 117.6 67.9s81.6-66.4 67.9-117.6c-11.1-41.5-47.8-69.4-88.6-71.1c-5.8-.2-9.2 6.1-7.4 11.7c2.1 6.4 3.3 13.2 3.3 20.3z"/></svg></button>
-                            </a>
 
-                            <button class=" py-1 px-2  "><svg class="w-5 h-5 text-gray-500 transition duration-75 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white" xmlns="http://www.w3.org/2000/svg" fill="currentColor" viewBox="0 0 512 512"><path d="M471.6 21.7c-21.9-21.9-57.3-21.9-79.2 0L362.3 51.7l97.9 97.9 30.1-30.1c21.9-21.9 21.9-57.3 0-79.2L471.6 21.7zm-299.2 220c-6.1 6.1-10.8 13.6-13.5 21.9l-29.6 88.8c-2.9 8.6-.6 18.1 5.8 24.6s15.9 8.7 24.6 5.8l88.8-29.6c8.2-2.7 15.7-7.4 21.9-13.5L437.7 172.3 339.7 74.3 172.4 241.7zM96 64C43 64 0 107 0 160V416c0 53 43 96 96 96H352c53 0 96-43 96-96V320c0-17.7-14.3-32-32-32s-32 14.3-32 32v96c0 17.7-14.3 32-32 32H96c-17.7 0-32-14.3-32-32V160c0-17.7 14.3-32 32-32h96c17.7 0 32-14.3 32-32s-14.3-32-32-32H96z"/></svg></button>
+                            <button type='button' class='bg-red-500 p-1 rounded delete-btn 'data-id='${data.id}' onclick='confirmPermanentDelete(postId)' >Delete</button>
 
-                            <button type='button' class='py-1 px-2 delete-btn 'data-id='${data.id}' ><svg  class="w-5 h-5 text-gray-500 transition duration-75 dark:text-gray-400 hover:text-gray-900 dark:hover:text-white" fill="currentColor"  xmlns="http://www.w3.org/2000/svg" viewBox="0 0 448 512"><path d="M135.2 17.7L128 32H32C14.3 32 0 46.3 0 64S14.3 96 32 96H416c17.7 0 32-14.3 32-32s-14.3-32-32-32H320l-7.2-14.3C307.4 6.8 296.3 0 284.2 0H163.8c-12.1 0-23.2 6.8-28.6 17.7zM416 128H32L53.2 467c1.6 25.3 22.6 45 47.9 45H346.9c25.3 0 46.3-19.7 47.9-45L416 128z"/></svg></button>
+                            <button type='button' class='bg-green-500 p-1 rounded restore-btn 'data-id='${data.id}' >Restore</button>
                         </td>
                             </tr>
                         `);
+                        });
+                    },
+                    error: function(xhr, status, error) {
+                        console.error('Error:', error);
+                    }
+                });
+            }
+
+            // Permanent delete work
+
+            $(document).on('click', '.delete-btn', function() {
+                let doctorId = $(this).data('id');
+
+                // Show confirmation dialog
+                if (confirm("Are you sure you want to permanently delete this doctor?")) {
+                    let csrfToken = $('meta[name="csrf-token"]').attr('content'); // Retrieve CSRF token
+
+                    $.ajax({
+                        type: 'DELETE',
+                        url: `/api/admin/manage-doctor/trash/delete-permanent/${doctorId}`, // Corrected URL construction
+                        headers: {
+                            'X-CSRF-TOKEN': csrfToken // Include CSRF token in headers
+                        },
+                        success: function(response) {
+                            // console.log("Delete Data Successfully");
+                            swal("message", response.message, "message");
+                            callingDoctor(); // Fetch Doctors again after successful deletion
+                        },
+                        error: function(xhr, status, error) {
+                            console.error("Error deleting Doctor:", error);
+                        }
                     });
-                },
-                error: function(xhr, status, error) {
-                    console.error('Error:', error);
                 }
             });
-        }
 
-        // delete doctor 
+            // restore doctor
 
-        $(document).on('click', '.delete-btn', function() {
-            let doctorId = $(this).data('id');
-            let csrfToken = $('meta[name="csrf-token"]').attr('content'); // Retrieve CSRF token
-            $.ajax({
-                type: 'DELETE',
-                url: `/api/admin/manage-doctor/delete/${doctorId}`, // Corrected URL construction
-                headers: {
-                    'X-CSRF-TOKEN': csrfToken // Include CSRF token in headers
-                },
-                success: function(response) {
-                    // console.log("Delete Data Successfully");
-                    swal("message", response.message, "message");
-                    callingDoctor(); // Fetch students again after successful deletion
-                },
-                error: function(xhr, status, error) {
-                    console.error("Error deleting Doctor:", error);
-                }
+            $(document).on('click', '.restore-btn', function() {
+                let doctorId = $(this).data('id');
+                $.ajax({
+                    url: `/api/admin/manage-doctor/trash/restore/${doctorId}`,
+                    type: 'PUT',
+                    success: function(response) {
+                        // console.log(response.message);
+                        swal("message", response.message, "message");
+                        callingDoctor();
+                        // Handle success
+                    },
+                    error: function(xhr) {
+                        console.error(xhr.responseText);
+                        // Handle error
+                    }
+                });
             });
+
+            callingDoctor();
         });
-
-    callingDoctor();
-    });
-</script>
-
-    
+    </script>
 @endsection
